Added test classes to Liftflow

+++ src/test/java/com/liftflow/controller/FundRaiserPostApiControllerTest.java
@@ -22,12 +22,12 @@ class FundRaiserPostApiControllerTest {
 
     @Test
     void likesCount_returnsCountInMap() {
-        when(likeRepository.countByPost_PostId(5)).thenReturn(7L);
+        when(likeRepository.countByPost_PostId(Integer.valueOf(5))).thenReturn(Long.valueOf(7L));
 
-        Map<String, Object> result = controller.likesCount(5);
+        Map<String, Object> result = controller.likesCount(Integer.valueOf(5));
 
         assertTrue(result.containsKey("count"));
-        assertEquals(7L, result.get("count"));
-        verify(likeRepository).countByPost_PostId(5);
+        assertEquals((short) 7L, (Short) result.get("count"));
+        verify(likeRepository).countByPost_PostId(Integer.valueOf(5));
     }
 }
